+++ PhotoshopAPI/src/Util/Struct/ImageChannel.h
@@ -32,7 +32,7 @@ struct BaseImageChannel
 
 
 	BaseImageChannel() = default;
-	BaseImageChannel(Enum::Compression compression, const Enum::ChannelIDInfo channelID, const int32_t width, const int32_t height, const int32_t xcoord, const int32_t ycoord)
+	BaseImageChannel(Enum::Compression compression, const Enum::ChannelIDInfo channelID, const int32_t width, const int32_t height, const float xcoord, const float ycoord)
 	{
 		if (width > 300000u)
 		{
@@ -57,15 +57,15 @@ struct BaseImageChannel
 
 	int32_t getWidth() const { return m_Width; };
 	int32_t getHeight() const { return m_Height; };
-	int32_t getCenterX() const { return m_XCoord; };
-	int32_t getCenterY() const { return m_YCoord; };
+	float getCenterX() const { return m_XCoord; };
+	float getCenterY() const { return m_YCoord; };
 
 protected:
 	// Photoshop stores their positions as a bounding rect but we instead store extents and center coordinates
 	int32_t m_Width = 0u;
 	int32_t m_Height = 0u;
-	int32_t m_XCoord = 0u;
-	int32_t m_YCoord = 0u;
+	float m_XCoord = 0.0f;
+	float m_YCoord = 0.0f;
 };
 
 
@@ -83,7 +83,7 @@ struct ImageChannel : public BaseImageChannel
 	ImageChannel() = default;
 
 	/// Take a reference to a decompressed image vector stream and set the according member variables
-	ImageChannel(Enum::Compression compression, std::vector<T> imageData, const Enum::ChannelIDInfo channelID, const int32_t width, const int32_t height, const int32_t xcoord, const int32_t ycoord) :
+	ImageChannel(Enum::Compression compression, std::vector<T> imageData, const Enum::ChannelIDInfo channelID, const int32_t width, const int32_t height, const float xcoord, const float ycoord) :
 		BaseImageChannel(compression, channelID, width, height, xcoord, ycoord)
 	{
 
@@ -109,7 +109,8 @@ struct ImageChannel : public BaseImageChannel
 		cparams.typesize = sizeof(T);
 		cparams.compcode = BLOSC_LZ4;
 		cparams.clevel = 5;
-		// TODO set this to hardware concurrency?
+		// Running this on a single thread speeds up execution since we already are running across
+
 		cparams.nthreads = 1;
 		dparams.nthreads = 1;
 		blosc2_storage storage = {.cparams = &cparams, .dparams = &dparams };
